+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.X.X](https://www.npmjs.com/package/aristid-ds/v/1.X.X) (2024-XX-XX)
+
+### 🐛 Bug Fixes
+* **AutoComplete:**
+  * Fix dropdown border when focused or when a status is set (@philippechevieux)
+
+* **Select:**
+  * Fix dropdown border when focused or when a status is set (@philippechevieux)
+
 ## [1.0.0](https://www.npmjs.com/package/aristid-ds/v/1.0.0) (2024-02-02)
 * **⚠️ Breaking change - Button:**  (@evoiron)
     * update color Management